# src/generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"DriverResponse": {
				"description": "Driver information response",
				"properties": {
					"id": {
						"type": "string",
						"description": "Unique driver identifier (UUID)"
					},
					"name": {
						"type": "string",
						"description": "Driver's full name"
					},
					"phone": {
						"type": "string",
						"description": "Phone number"
					},
					"vehicleType": {
						"type": "string",
						"description": "Vehicle type: 'bike', 'scooter', or 'car'"
					},
					"status": {
						"type": "string",
						"description": "Current driver status: 'offline', 'available', 'en_route_pickup', 'at_pickup', 'en_route_delivery', 'at_delivery'"
					},
					"maxOrders": {
						"type": "number",
						"format": "double",
						"description": "Maximum concurrent orders"
					}
				},
				"required": [
					"id",
					"name",
					"phone",
					"vehicleType",
					"status",
					"maxOrders"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"id": "550e8400-e29b-41d4-a716-446655440000",
					"name": "John Doe",
					"phone": "+1234567890",
					"vehicleType": "bike",
					"status": "available",
					"maxOrders": 3
				}
			},
			"CreateDriverRequest": {
				"description": "Request body for creating a new driver",
				"properties": {
					"name": {
						"type": "string",
						"description": "Driver's full name"
					},
					"phone": {
						"type": "string",
						"description": "Phone number with country code (e.g., +1234567890)"
					},
					"vehicleType": {
						"type": "string",
						"description": "Type of vehicle: 'bike', 'scooter', or 'car'"
					},
					"maxOrders": {
						"type": "number",
						"format": "double",
						"description": "Maximum number of concurrent orders (default: 3)"
					},
					"initialLocation": {
						"properties": {
							"lng": {
								"type": "number",
								"format": "double",
								"description": "Longitude coordinate"
							},
							"lat": {
								"type": "number",
								"format": "double",
								"description": "Latitude coordinate"
							}
						},
						"required": [
							"lng",
							"lat"
						],
						"type": "object",
						"description": "Initial GPS location of the driver"
					}
				},
				"required": [
					"name",
					"phone",
					"vehicleType"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "John Doe",
					"phone": "+1234567890",
					"vehicleType": "bike",
					"maxOrders": 3,
					"initialLocation": {
						"lat": 37.7749,
						"lng": -122.4194
					}
				}
			},
			"UpdateLocationRequest": {
				"description": "Request body for updating driver's GPS location",
				"properties": {
					"lat": {
						"type": "number",
						"format": "double",
						"description": "Latitude coordinate (-90 to 90)"
					},
					"lng": {
						"type": "number",
						"format": "double",
						"description": "Longitude coordinate (-180 to 180)"
					},
					"heading": {
						"type": "number",
						"format": "double",
						"description": "Compass heading in degrees (0-360, optional)"
					},
					"speed": {
						"type": "number",
						"format": "double",
						"description": "Speed in km/h (optional)"
					}
				},
				"required": [
					"lat",
					"lng"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"lat": 37.7749,
					"lng": -122.4194,
					"heading": 90,
					"speed": 15.5
				}
			},
			"UpdateStatusRequest": {
				"description": "Request body for updating driver status",
				"properties": {
					"status": {
						"type": "string",
						"description": "New driver status: 'offline', 'available', 'en_route_pickup', 'at_pickup', 'en_route_delivery', 'at_delivery'"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": "available"
				}
			},
			"RouteAssignment": {
				"description": "Route assignment information for an order",
				"properties": {
					"assignmentId": {
						"type": "string"
					},
					"orderId": {
						"type": "string"
					},
					"sequence": {
						"type": "number",
						"format": "double"
					},
					"estimatedPickup": {
						"type": "string",
						"format": "date-time"
					},
					"estimatedDelivery": {
						"type": "string",
						"format": "date-time"
					},
					"actualPickup": {
						"type": "string",
						"format": "date-time"
					},
					"actualDelivery": {
						"type": "string",
						"format": "date-time"
					},
					"pickupLat": {
						"type": "number",
						"format": "double"
					},
					"pickupLng": {
						"type": "number",
						"format": "double"
					},
					"pickupAddress": {
						"type": "string"
					},
					"dropoffLat": {
						"type": "number",
						"format": "double"
					},
					"dropoffLng": {
						"type": "number",
						"format": "double"
					},
					"dropoffAddress": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"assignmentId",
					"orderId",
					"sequence",
					"estimatedPickup",
					"estimatedDelivery",
					"pickupLat",
					"pickupLng",
					"pickupAddress",
					"dropoffLat",
					"dropoffLng",
					"dropoffAddress",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteStop": {
				"description": "Individual stop in the route (pickup or delivery)",
				"properties": {
					"stopIndex": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"pickup",
							"delivery"
						]
					},
					"lat": {
						"type": "number",
						"format": "double"
					},
					"lng": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"distanceFromPrevious": {
						"type": "number",
						"format": "double"
					},
					"durationFromPrevious": {
						"type": "number",
						"format": "double"
					},
					"durationFromPreviousFormatted": {
						"type": "string"
					},
					"estimatedArrival": {
						"type": "string",
						"format": "date-time"
					},
					"cumulativeDistance": {
						"type": "number",
						"format": "double"
					},
					"cumulativeTime": {
						"type": "number",
						"format": "double"
					},
					"cumulativeTimeFormatted": {
						"type": "string"
					}
				},
				"required": [
					"stopIndex",
					"orderId",
					"type",
					"lat",
					"lng",
					"address",
					"distanceFromPrevious",
					"durationFromPrevious",
					"durationFromPreviousFormatted",
					"estimatedArrival",
					"cumulativeDistance",
					"cumulativeTime",
					"cumulativeTimeFormatted"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DriverRouteResponse": {
				"description": "Driver route response with optimized stop sequence",
				"properties": {
					"driverId": {
						"type": "string"
					},
					"driverName": {
						"type": "string"
					},
					"driverStatus": {
						"type": "string"
					},
					"totalAssignments": {
						"type": "number",
						"format": "double"
					},
					"totalStops": {
						"type": "number",
						"format": "double"
					},
					"currentLoad": {
						"type": "number",
						"format": "double"
					},
					"maxCapacity": {
						"type": "number",
						"format": "double"
					},
					"assignments": {
						"items": {
							"$ref": "#/components/schemas/RouteAssignment"
						},
						"type": "array"
					},
					"stops": {
						"items": {
							"$ref": "#/components/schemas/RouteStop"
						},
						"type": "array"
					},
					"totalDistance": {
						"type": "number",
						"format": "double"
					},
					"estimatedDuration": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"driverId",
					"driverName",
					"driverStatus",
					"totalAssignments",
					"totalStops",
					"currentLoad",
					"maxCapacity",
					"assignments",
					"stops",
					"totalDistance",
					"estimatedDuration"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "EdgeRun API",
		"version": "1.0.0",
		"description": "EdgeRun Last-Mile Delivery API - MVP for food delivery driver-order matching",
		"license": {
			"name": {
				"name": "MIT"
			}
		},
		"contact": {
			"name": "EdgeRun Support"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/drivers": {
			"post": {
				"operationId": "CreateDriver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DriverResponse"
								}
							}
						}
					},
					"201": {
						"description": "Driver created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DriverResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDriverRequest"
							}
						}
					}
				}
			}
		},
		"/drivers/{id}/location": {
			"post": {
				"operationId": "UpdateLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLocationRequest"
							}
						}
					}
				}
			}
		},
		"/drivers/{id}/status": {
			"patch": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DriverResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStatusRequest"
							}
						}
					}
				}
			}
		},
		"/drivers/{id}": {
			"get": {
				"operationId": "GetDriver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DriverResponse"
								}
							}
						}
					},
					"404": {
						"description": "Driver not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/drivers/{id}/assignments/offered": {
			"get": {
				"operationId": "GetOfferedAssignments",
				"responses": {
					"200": {
						"description": "Array of offered assignments",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RouteAssignment"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Driver not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get offered assignments for a driver (pending response)\nReturns all assignments with OFFERED status",
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Driver ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/drivers/assignments/{id}/accept": {
			"post": {
				"operationId": "AcceptAssignment",
				"responses": {
					"200": {
						"description": "Success message with updated assignment",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"assignment": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"assignment",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid state transition",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Assignment not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Accept an offered assignment",
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Assignment ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/drivers/assignments/{id}/reject": {
			"post": {
				"operationId": "RejectAssignment",
				"responses": {
					"200": {
						"description": "Success message",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid state transition",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Assignment not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Reject an offered assignment",
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Assignment ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Rejection reason (optional)",
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"reason": {
										"type": "string"
									}
								},
								"type": "object",
								"description": "Rejection reason (optional)"
							}
						}
					}
				}
			}
		},
		"/drivers/{id}/route": {
			"get": {
				"operationId": "GetDriverRoute",
				"responses": {
					"200": {
						"description": "DriverRouteResponse with optimized stop sequence",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DriverRouteResponse"
								}
							}
						}
					},
					"404": {
						"description": "Driver not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get the current delivery route for a driver\nReturns optimized stop sequence (pickups & deliveries) respecting capacity constraints",
				"tags": [
					"Drivers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Driver ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}